feat: hairstyels and ability to tie hairsyle
Allows you to braid your hair into an alternate version of a hairstyle.

Click the middle mouse button on your head.

To change the hairstyle to a different style - specify the type
of hairstyle in the `tied_type` variable

Don't forget to do the same with the second style so you can
change the hairstyle back.